Create 4H lockup style and ability to generate
Extension wants to have a style with 4H logo included so here it is.
This commit gives the generator the ability to create the lockup,
zip up the files, the whole works. The paths are included, and the
green color, for reference, is 129a63 in RGB, Pantone 347U for Pantone.

Progress on completing Issue #1